requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.27.0
google-search-results>=2.4.2
feedparser>=6.0.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
aiosqlite>=0.19.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.28.0
